Replace theme registry and menu alters with a form_alter
Overriding the theme registry and menu is a slightly awkward and non-standard way of replacing a form in Drupal so I replaced that with a form_alter which replaces the contents of the user_login and user_login_block forms with a Auth0 widget.

This is simlper conceptually and does allows the core /user redirect functionality and permissions to stay intact.